Dedupe __strToRunes and __runesToStr
They were defined in tests and in trie archive.
Plan to add more rune utils, so putting them in their own compilation unit
